Added very basic searching using strstr(3).
This is a good start, but currently the output from .local.desc files
doesn't make much sense, as it starts printing from the beginning of a
substring, which may be in the middle of a line. TODO: implement an
algorithm which can find the first newline character before the
beginning of the substring match, and print from there until the next
newline.