Completed documentation for functions; small fix.
See the various TODOs in the code. A decent alpha release could come
soon, however I want to do some more rigorous testing first.  Error-
reporting also needs to be improved, allowing for error messages to
print the file to which they take issue.

Command-line options are on the horizon; this should probably exist as
the (one) global variable: an integer with various bits set for an
option, defined in an enum as various powers of two, i.e.

        #define SET_BIT(val, n) ( val |= n )
        #define CHK_BIT(val, n) ( val &  n )

        enum opt_positions {
                OPT_1 = 1,
                OPT_2 = 2,
                OPT_3 = 4,
                OPT_4 = 8,
                /* etc. */
        }

        uint8_t options = 0; /* should be fixed-size; c.f. <stdint.h> */