MINOR: include website in the version information.
The argument-processor should be improved before the first Alpha
release, such that "-a -b" == "-ab" == "-ba" ("-bb" or "-aa" should
still cause a double-definition error). It was going to be pushed back
until the post-Alpha/Beta release, although it seems to be simpler than
originally anticipated. TODO: implement this and amend the README
accordingly; man page ?