Generalised the reporter to include warnings.
The appropriate symbols have been renamed to error* to info*, and a
'severity' parameter was introduced; this purely changes the nature of
the displayed message; this only changes the text displayed, and not the
actual behaviour or status code of the program.

GOTO statements were used in print_info purely due to the awkward
tabulations and line-breaks which could ensure, if the print statements
were indented with sixteen additional spaces.