Completed the Two-Way searching algorithm.
It will stay separate from the primary codebase for a while, for the
reasons described atop twoway.c. It's a shame I can't just define
CANON_ELEMENT on a decompiled glibc and call strstr(3), (exactly) as
strcasestr does ! It is a huge shame that the latter function is not
defined as part of the standard library.