Temporary work-around for multi-buffer colouring.
See the (urgent) TODO in `print_coloured_result`, although it seems like
it might be quite a difficult fix. Perhaps, if the buffer is truncated
and the value returned by sep{1,2}_idx is unrealistic/unaddressable
(i.e., over BUFFER_SZ), use `strchr` and `strstr` on the rescue buffer
(of size SBUF_SZ) instead.